Move retry functions to their own module
