+++ flocker/testtools/__init__.py
@@ -14,12 +14,10 @@ import os
 import pwd
 from collections import namedtuple
 from contextlib import contextmanager
-from itertools import repeat
 from random import randrange
 import shutil
-from functools import wraps, partial
+from functools import wraps
 from unittest import skipIf, skipUnless
-from inspect import getfile, getsourcelines
 from StringIO import StringIO
 from subprocess import PIPE, STDOUT, CalledProcessError, Popen
 
@@ -28,7 +26,7 @@ from bitmath import GiB, MiB
 from pyrsistent import PRecord, field
 
 from docker import Client as DockerClient
-from eliot import ActionType, Message, MessageType, start_action, fields, Field
+from eliot import ActionType, Message, MessageType, start_action, fields
 from eliot.twisted import DeferredContext
 
 from zope.interface import implementer
@@ -38,9 +36,8 @@ from twisted.internet.interfaces import (
     IProcessTransport, IReactorProcess, IReactorCore,
 )
 from twisted.python.filepath import FilePath, Permissions
-from twisted.python.reflect import prefixedMethodNames, safe_repr
-from twisted.internet.task import Clock, deferLater
-from twisted.internet.defer import maybeDeferred, Deferred
+from twisted.internet.task import Clock
+from twisted.internet.defer import Deferred
 from twisted.internet.error import ConnectionDone
 from twisted.internet import reactor
 from twisted.trial.unittest import SynchronousTestCase, SkipTest
@@ -193,137 +190,6 @@ def assert_not_equal_comparison(case, a, b):
     if messages:
         case.fail(
             "Expected a and b to be not-equal: " + "; ".join(messages))
-
-
-def function_serializer(function):
-    """
-    Serialize the given function for logging by eliot.
-
-    :param function: Function to serialize.
-
-    :return: Serialized version of function for inclusion in logs.
-    """
-    try:
-        return {
-            "function": str(function),
-            "file": getfile(function),
-            "line": getsourcelines(function)[1]
-        }
-    except IOError:
-        # One debugging method involves changing .py files and is incompatible
-        # with inspecting the source.
-        return {
-            "function": str(function),
-        }
-
-
-class LoopExceeded(Exception):
-    """
-    Raised when ``loop_until`` looped too many times.
-    """
-
-    def __init__(self, predicate, last_result):
-        super(LoopExceeded, self).__init__(
-            '%r never True in loop_until, last result: %r'
-            % (predicate, last_result))
-
-
-LOOP_UNTIL_ACTION = ActionType(
-    action_type="flocker:testtools:loop_until",
-    startFields=[Field("predicate", function_serializer)],
-    successFields=[Field("result", serializer=safe_repr)],
-    description="Looping until predicate is true.")
-
-LOOP_UNTIL_ITERATION_MESSAGE = MessageType(
-    message_type="flocker:testtools:loop_until:iteration",
-    fields=[Field("result", serializer=safe_repr)],
-    description="Predicate failed, trying again.")
-
-
-def loop_until(predicate, reactor=reactor, steps=None):
-    """Repeatedly call ``predicate``, until it returns something ``Truthy``.
-
-    :param predicate: Callable returning termination condition.
-    :type predicate: 0-argument callable returning a Deferred.
-
-    :param reactor: The reactor implementation to use to delay.
-    :type reactor: ``IReactorTime``.
-
-    :param steps: An iterable of delay intervals, measured in seconds.
-        If not provided, will default to retrying every 0.1 seconds forever.
-
-    :raise LoopExceeded: If given a finite sequence of steps, and we exhaust
-        that sequence waiting for predicate to be truthy.
-
-    :return: A ``Deferred`` firing with the first ``Truthy`` response from
-        ``predicate``.
-    """
-    if steps is None:
-        steps = repeat(0.1)
-    else:
-        steps = iter(steps)
-
-    action = LOOP_UNTIL_ACTION(predicate=predicate)
-
-    d = action.run(DeferredContext, maybeDeferred(action.run, predicate))
-
-    def loop(result):
-        if not result:
-            LOOP_UNTIL_ITERATION_MESSAGE(
-                result=result
-            ).write()
-            try:
-                delay = steps.next()
-            except StopIteration:
-                raise LoopExceeded(predicate, result)
-            d = deferLater(reactor, delay, action.run, predicate)
-            d.addCallback(partial(action.run, loop))
-            return d
-        action.addSuccessFields(result=result)
-        return result
-    d.addCallback(loop)
-    return d.addActionFinish()
-
-
-def retry_failure(function, expected=None, reactor=reactor, steps=None):
-    """
-    Retry ``function`` until it returns successfully.
-
-    If it raises one of the expected exceptions, then retry.
-
-    :param callable function: A callable that returns a value.
-    :param expected: Iterable of exceptions that trigger a retry. Passed
-        through to ``Failure.check``.
-    :param reactor reactor: The reactor implementation to use to delay.
-    :param [float] steps: An iterable of delay intervals, measured in seconds.
-        If not provided, will default to retrying every 0.1 seconds.
-
-    :return: A ``Deferred`` that fires with the first successful return value
-        of ``function``.
-    """
-    if steps is None:
-        steps = repeat(0.1)
-    else:
-        steps = iter(steps)
-
-    d = maybeDeferred(function)
-
-    def loop(failure):
-        if expected and not failure.check(*expected):
-            return failure
-
-        try:
-            interval = steps.next()
-        except StopIteration:
-            return failure
-
-        d = deferLater(reactor, interval, function)
-        d.addErrback(loop)
-        return d
-
-    d.addErrback(loop)
-
-    return d
 
 
 def random_name(case):
